+++ Project/react-web/package.json
@@ -16,7 +16,8 @@
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.1",
-    "styled-components": "^5.3.10"
+    "styled-components": "^5.3.10",
+    "styled-reset": "^4.4.7"
   },
   "scripts": {
     "start": "react-scripts start",
