+++ Project/react-web/src/App.js
@@ -1,5 +1,8 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
+import { ThemeProvider } from "styled-components";
+import GlobalStyle from "./styles/GlobalStyle";
+import theme from "./styles/theme";
 
 import SignIn from "./pages/SignIn";
 import SignUp from "./pages/SignUp";
@@ -8,12 +11,15 @@ import Application from "./pages/Application";
 function App() {
   return (
     <div className="App">
-      <Routes>
-        <Route path="/" element={<SignIn />}></Route>
-        <Route path="signin/" element={<SignIn />}></Route>
-        <Route path="/signup" element={<SignUp />}></Route>
-        <Route path="/application" element={<Application />}></Route>
-      </Routes>
+      <ThemeProvider theme={theme}>
+        <GlobalStyle />
+        <Routes>
+          <Route path="/" element={<SignIn />}></Route>
+          <Route path="signin/" element={<SignIn />}></Route>
+          <Route path="/signup" element={<SignUp />}></Route>
+          <Route path="/application" element={<Application />}></Route>
+        </Routes>
+      </ThemeProvider>
     </div>
   );
 }
