+++ Project/react-web/src/styles/GlobalStyle.js
@@ -1,0 +1,65 @@
+import { createGlobalStyle } from "styled-components";
+import reset from "styled-reset";
+
+const GlobalStyle = createGlobalStyle`
+  ${reset};
+  @font-face {
+    font-family: 'Pretendard-Regular';
+    src: url('https://cdn.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
+    font-weight: 400;
+    font-style: normal;
+  }
+
+  html,
+  body {
+    width: 100%;
+    height: 100%;
+  }
+
+  :root{
+    --vh: 100%;
+
+  }
+  
+  #root {
+    margin: 0 auto;
+  }
+  
+  html {
+    font-size: 62.5%;
+  }
+  
+  * {
+    box-sizing: border-box;
+  }
+  
+  body, button {
+    font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans,
+    Helvetica Neue, sans-serif;
+  }
+  
+  button {
+    cursor: pointer;
+    border: none;
+    outline: none;
+    background-color: transparent;
+    -webkit-tap-highlight-color : transparent;
+  }
+
+  input {
+    outline: none;
+  }
+  
+  a, a:visited {
+    text-decoration: none;
+    color: black;
+  }
+
+  input[type='number']::-webkit-outer-spin-button,
+  input[type='number']::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+  }
+`;
+
+export default GlobalStyle;
